add one-import usage for sttp (https://github.com/softwaremill/sttp/issues/121)
