+++ core/jvm/src/main/scala/com/softwaremill/sttp/simpleApi.scala
@@ -1,0 +1,5 @@
+package com.softwaremill.sttp
+
+object simpleApi extends SttpApi {
+  implicit lazy val backend = HttpURLConnectionBackend()
+}
